+++ enterprise/workers/socket/package.json
@@ -15,6 +15,7 @@
 	"dependencies": {
 		"@tsndr/cloudflare-worker-jwt": "^3.2.0",
 		"@types/jsonwebtoken": "^8.5.9",
+		"hono": "^4.8.2",
 		"jsonwebtoken": "9.0.0",
 		"ws": "^8.18.0"
 	}
