+++ libs/application-generic/src/services/index.ts
@@ -27,3 +27,4 @@ export * from './cron';
 export * from './content.service';
 export * from './sanitize/sanitizer.service';
 export * from './sanitize/sanitizer-v0.service';
+export * from './socket-worker';
