+++ libs/application-generic/src/services/socket-worker/index.ts
@@ -1,0 +1,1 @@
+export { SocketWorkerService } from './socket-worker.service';
